Show and Create post page
Created the show and create page. Created a PostLinks component where I put a link to the create post page. Created the Message component. Added a Add and Fail SVG icon component to be in the message component. Created TextArea component. Updated the auth hook to return csfr token and the mutate function.